refactor PrismaDBExceptionHandler
